requirements.txt
requests
rich
python-dateutil